Major update of the install.py script.
This is now completely refactored to remove all config file parsing
and file downloading functionality. Instead, it uses the common.py
and configfile.py modules for this functionality.

It's still using the old optparse parser. This will need to be updated
once Gabriel checks in the new argparse parser.

Added try/catch statements to main() to catch runtime errors and ^C.

Added 'empty' and 'changed' properties to ConfigFile