DRTVWR-418: Allow autobuild source_environment to set arbitrary vars.
Until now we've had specific placeholders for specific environment-variable
assignments in the autobuild source_environment template string. That meant
that adding a new variable assignment required adding both a dict entry and a
new template placeholder for it. Forgetting the former caused an error;
forgetting the latter (easy to do!) silently failed, omitting the new
assignment.

Instead, replace all the specific variable-assignment placeholders with a
single generic one, with logic to massage all the entries in a particular dict
into a sequence of variable-assignment statements to replace that placeholder.
In practice we use two dicts: one for assignments that should also be 'export'
statements, the other for assignments that can be local to the running bash
process.

Migrate a couple of the recent Windows-specific AUTOBUILD_xxx variables to use
this new mechanism, so they're set whether or not the bash script invokes
load_vsvars. That leaves load_vsvars setting only variables actually set by
the vsvarsall.bat script, as originally intended.