SL-525: Fix test_install localhost 503 problem.
Though it's often useful to have an http_proxy environment variable set,
that's disastrous when setting up a temporary localhost server as we do in
this test script. You DO NOT WANT Python to proxy a localhost:port request
through some remote server!

Also eliminate the find-free-port retry loop: this is unnecessary when you
specify port 0 (allowing the OS to select the port) and then capture the
assigned port number.