SL-525: Incorrect add_argument('--dry-run') call broke unit tests.
Don't pass ('-n', '--dry-run') as a tuple, just as individual arguments.
ArgumentParser.add_argument() figures it out from there.

Passing a tuple confused add_argument(), which stored the tuple as the 'dest'
for the action, which caused a TypeError deep in parse_args() when the tuple
got passed to hasattr() as the attribute name.