SL-525: Only expand autobuild.xml $vars for configure and build.
It turns out to be a spectacularly bad idea to expand $vars in autobuild.xml
data for any command that is going to resave autobuild.xml!

Add logic to ConfigurationDescription to forbid resaving if $vars have been
expanded.

Add logic to each of autobuild_tool_configure and autobuild_tool_build to
expand $vars there.

Add tests verifying each of the above changes.