Fix SCM version discovery from untracked dir
Autobuild is commonly run from a staging directory not checked into
root. When this is done, SCM version discovery erronously reports that
the current version is dirty. This is due to git assuming the CWD is the
repository's root when using `--git-dir`.

Fix: Set `-C` to specify the repository root and avoid
incorrectly reporting that the repository is dirty.