SL-17226: Remove source_environment assumption about buildscripts.
autobuild_tool_source_environment.py checks to see whether the autobuild repo
has been checked out beside one called 'buildscripts' to look for a
get_params.py script.

get_params has never actually been used - it has remained a TODO since this
subcommand was first introduced in April 2010. Let's concede that if we
haven't felt the need over the last 12 years, we're probably not going to.

Moreover, for years now our builds have used 'sling-buildscripts' instead of
'buildscripts', so even on a CI host the logic wouldn't work.

The one effect of this code is a land mine: if autobuild is run on a machine
on which a developer just HAPPENS to have a clone of the old 'buildscripts'
repo sitting beside the autobuild repo, it tries to import a very old Python 2
module and crashes autobuild.

It's way past time to remove the get_params stuff.