Add support for downloading packages from private sources
This changeset provides support for downloading autobuild packages from
restricted sources such as private repositories using GitHub release
artifacts or GitLab generic packages.

The way this works is by the addition of an optional installable
parameter: `creds`, which may be specified when adding or editing
dependencies:

```
autobuild installables edit my-pkg creds=github url=...
```

This property is used to inform autobuild that it should set a HTTP
authorization header specific to the provider (github, gitlab) when
downloading the respective artifact.

In the future, this could be enhanced to include more providers or
authorization schemes. The `autobuild install` command could also be
enhanced with equivilent kwargs `--github-token` and `--gitlab-token`.